Refactor property tests
