+++ src/db_iter.rs
@@ -1,10 +1,10 @@
+use crate::bytes_range::BytesRange;
 use crate::config::DbRecord;
 use crate::db_state::{DbStateSnapshot, SsTableHandle};
 use crate::error::SlateDBError;
 use crate::iter::KeyValueIterator;
 use crate::mem_table::VecDequeKeyValueIterator;
 use crate::merge_iterator::{MergeIterator, TwoMergeIterator};
-use crate::range_util::BytesRange;
 use crate::sorted_run_iterator::SortedRunIterator;
 use crate::sst_iter::SstIterator;
 use bytes::Bytes;
