+++ src/mem_table.rs
@@ -8,11 +8,11 @@ use crossbeam_skiplist::map::Range;
 use crossbeam_skiplist::SkipMap;
 use tokio::sync::watch;
 
+use crate::bytes_range::BytesRange;
 use crate::db_iter::SeekToKey;
 use crate::error::SlateDBError;
 use crate::iter::KeyValueIterator;
 use crate::merge_iterator::MergeIterator;
-use crate::range_util::BytesRange;
 use crate::types::{RowAttributes, RowEntry, ValueDeletable};
 
 pub(crate) struct KVTable {
