+++ src/sst_iter.rs
@@ -6,11 +6,11 @@ use std::ops::Bound::{Excluded, Included, Unbounded};
 use std::sync::Arc;
 use tokio::task::JoinHandle;
 
+use crate::bytes_range::BytesRange;
 use crate::db_iter::SeekToKey;
 use crate::db_state::SsTableHandle;
 use crate::error::SlateDBError;
 use crate::flatbuffer_types::{SsTableIndex, SsTableIndexOwned};
-use crate::range_util::BytesRange;
 use crate::{
     block::Block, block_iterator::BlockIterator, iter::KeyValueIterator, tablestore::TableStore,
     types::RowEntry,
